config/requirements.txt
flask==2.0.1
flask_httpauth==4.4.0
pyAesCrypt==6.0.0
requests==2.25.1
pyyaml==5.4.1
